Verify a card id exists before update
There is a case where the issue query returns an issue without a trello
card id attached to it which will crash. Going to detect first if the
attribute exists and avoid execution. If we have an alternate case here,
we can make an else or elif as part of this.